design: 프로필 수정 UI 수정
